Refactor Vulkan rendering responsibilities
