+++ frame/vulkan/CMakeLists.txt
@@ -17,8 +17,14 @@ set(FRAME_VULKAN_SOURCES
   mesh_resources.h
   material.cpp
   material.h
+  output_image_resources.cpp
+  output_image_resources.h
+  pipeline_resources.cpp
+  pipeline_resources.h
   program.cpp
   program.h
+  renderer.cpp
+  renderer.h
   command_queue.cpp
   command_queue.h
   gpu_memory_manager.cpp
